+++ src/TransmissionExtras.Server/EventIds.cs
@@ -20,6 +20,7 @@ public static class EventIds
         public const int RemovingTorrentsFailed = ClassId + 1;
         public const int DryRun = ClassId + 2;
         public const int RemovedTorrent = ClassId + 3;
+        public const int RemovingTorrentsJobStopping = ClassId + 4;
     }
 
     public static class VerifyTorrentsJob
@@ -30,5 +31,6 @@ public static class EventIds
         public const int VerifyingTorrentsFailed = ClassId + 1;
         public const int DryRun = ClassId + 2;
         public const int VerifyingTorrent = ClassId + 3;
+        public const int VerifyTorrentsJobStopping = ClassId + 4;
     }
 }